boxity_backend/requirements.txt
Flask==3.0.3
gunicorn
google-generativeai==0.8.2
requests==2.32.3
Pillow==10.4.0
flask-cors==5.0.0
jsonschema==4.23.0
opencv-python-headless==4.10.0.84
numpy==2.1.2
torch
torchvision
opencv-python
scikit-image
pillow
pandas
tqdm
fastapi
uvicorn[standard]
scikit-learn
python-dotenv==1.0.1
PyJWT==2.8.0
cryptography==42.0.0
